+++ changelog.md
@@ -1,3 +1,8 @@
+2.1.0
+  - added settings to disable SL and/or TP copying
+  - added settings to specify max and min trade volume
+  - added settings to configure trade size scaling
+
 1.1.1
   - handle TooManyRequestsException in HTTP client